+++ src/schema/slack/types/custom/form_input.ts
@@ -1,6 +1,6 @@
 import SchemaTypes from "../../../schema_types.ts";
 import { DefineType } from "../../../../types/mod.ts";
-import { DefineParameter } from "../../../../types/objects.ts";
+import { DefineParameter } from "../../../../types/define_parameter.ts";
 
 const FormInputObject = DefineParameter({
   description: "Input fields to be shown on the form",
